requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Pillow==10.2.0
email-validator==2.1.0.post1
bleach==6.1.0
Werkzeug==3.0.1
WTForms==3.1.2 